Refactor: profile statistics and post age
