+++ package.json
@@ -16,6 +16,7 @@
     "@react-navigation/native-stack": "^6.9.17",
     "@tanstack/react-query": "^5.13.4",
     "@types/jest": "^29.5.11",
+    "date-fns": "^3.2.0",
     "firebase": "^10.7.1",
     "formik": "^2.4.5",
     "react": "18.2.0",
